requirements.txt
pandas
faker
temporalio
pymongo
python-dotenv